+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Fixed
+* Fixed an issue where some animation files failed to load with `load_animations` for xc3_model.
+
 ## 0.8.0 - 2024-04-27
 ### Added
 * Added `Animation::skinning_transforms` and `Animation::model_space_transforms` to xc3_model.